Added basic Towny integration.
Add two checks and config options for integration with Towny. The first
one checks to see if the where the player is placing a bookshelf is in
their town if that check is enabled. The second check is to see if they
own the plot where the block is being placed at if that check is
enabled.